fix: update get_manager_usernames imports after refactoring
Update remaining imports of get_manager_usernames from
vibe3.config.orchestra_config to vibe3.services.orchestra_helpers
after the config layering refactoring.

Files affected:
- src/vibe3/domain/dispatch_coordinator.py
- src/vibe3/server/app.py
