+++ src/vibe3/domain/dispatch_coordinator.py
@@ -20,7 +20,6 @@ from typing import TYPE_CHECKING, Callable, cast
 from loguru import logger
 
 from vibe3.clients.github_client import GitHubClient
-from vibe3.config.orchestra_config import get_manager_usernames
 from vibe3.domain import publish
 from vibe3.domain.protocols.dispatch_protocols import (
     IssueCollectionServiceProtocol,
@@ -57,6 +56,7 @@ from vibe3.services.label_utils import (
     clean_old_state_labels,
     should_skip_from_queue,
 )
+from vibe3.services.orchestra_helpers import get_manager_usernames
 
 if TYPE_CHECKING:
     from vibe3.clients.sqlite_client import SQLiteClient
